+++ backend/canisters/local_user_index/impl/src/bots.rs
@@ -77,7 +77,7 @@ fn extract_access_context_from_jwt(jwt: &str, bot: &User, state: &mut RuntimeSta
             (
                 exp,
                 scope,
-                BotInitiator::ApiKeyPermissions(bot_action_claims.granted_permissions),
+                BotInitiator::ApiKeyPermissions(bot_action_claims.granted_permissions.into()),
                 bot_action_claims.bot,
             )
         } else {