+++ backend/libraries/types/src/bots.rs
@@ -251,6 +251,14 @@ pub struct BotCommand {
     pub name: String,
     pub args: Vec<BotCommandArg>,
     pub initiator: UserId,
+    pub meta: Option<BotCommandMeta>,
+}
+
+#[ts_export]
+#[derive(CandidType, Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub struct BotCommandMeta {
+    pub timezone: String, // IANA timezone e.g. "Europe/London"
+    pub language: String, // The language selected in OpenChat e.g. "en"
 }
 
 #[ts_export]
@@ -372,6 +380,12 @@ pub struct EncodedBotPermissions {
     message: Option<u32>,
 }
 
+impl From<BotPermissions> for EncodedBotPermissions {
+    fn from(permissions: BotPermissions) -> Self {
+        EncodedBotPermissions::from(&permissions)
+    }
+}
+
 impl From<&BotPermissions> for EncodedBotPermissions {
     fn from(permissions: &BotPermissions) -> Self {
         fn encode<T: Into<u8> + Copy>(field: &HashSet<T>) -> Option<u32> {
@@ -387,6 +401,12 @@ impl From<&BotPermissions> for EncodedBotPermissions {
             chat: encode(&permissions.chat),
             message: encode(&permissions.message),
         }
+    }
+}
+
+impl From<EncodedBotPermissions> for BotPermissions {
+    fn from(permissions: EncodedBotPermissions) -> Self {
+        BotPermissions::from(&permissions)
     }
 }