+++ backend/libraries/utils/src/base64.rs
@@ -7,11 +7,11 @@ where
     T: Serialize,
 {
     let json = serde_json::to_vec(value).unwrap();
-    base64::engine::general_purpose::STANDARD_NO_PAD.encode(json)
+    base64::engine::general_purpose::STANDARD.encode(json)
 }
 
 pub fn to_value<T: DeserializeOwned>(text: &str) -> Result<T, Box<dyn std::error::Error + Sync + Send>> {
-    let json = base64::engine::general_purpose::STANDARD_NO_PAD.decode(text)?;
+    let json = base64::engine::general_purpose::STANDARD.decode(text)?;
     let result = serde_json::from_slice(&json)?;
     Ok(result)
 }